sets.py: Adjust basis component to fit with new function in utils.py. I don't think I like the way I have the util built, but it will do for now.
inputs.py: Fixing up the from_file functionality. Ran into a problem
with "KIND" sections once again. May need to rethink the structure
in the future.

utils.py: a new util for getting basis and pseudopotential for now.
It will need to be overhauled anyway when we generalize it, but this
hacked-together-version is a least more useful for our current benchmarking

outputs.py: First pass at dealing with the output when OT is not used.
MO eigenvalues and band gap are a nightmare to get to behave properly when
OT is not used. Hopefully we can avoid using it too much, but for general
use it should be fixed up.