Remove rounding during FEFF writing (#3345)
* Remove rounding during FEFF writing

pymatgen/io/feff/inputs.py lines 504-506 rounded the FEFF geometry
to 6 decimal places by default. This also occurs in 518-520.
Other writers (such as the one for EXCITING) do not appear to do this.
This commit makes a simple change and removes the ":f" from the string
representation in the lines formatting. Thus FEFF clusters will now
be saved with default precision (or whatever self._cluster uses).

* list.append -> concat

---------

Co-authored-by: Janosh Riebesell <janosh.riebesell@gmail.com>